+++ server/src/controllers/auth.controller.ts
