--- server/src/controllers/index.ts
@@ -1,27 +1,0 @@
-export {
-} from "./college-student";
-
-export { getStudentInformationBySchoolId, createStudentInformation } from './student'
-
-export {
-  getCollegeSubjects,
-  getCollegeSubjectById,
-  getCollegeStudentsBySubjectId,
-  getCollegeSectionsBySubjectId,
-  getCollegeInstructorBySubjectId,
-} from "./college-subject";
-
-export {
-  createInstructor
-} from "./instructor";
-
-export { verifyMetaMessengerWebhook, handleReferralEvent } from "./webhook";
-
-export { handleGoogleCallback, validateOAuthToken } from "./auth";
-
-export { getSelf, createUser } from './user'
-
-export { getStrands, createStrand } from './strand'
-
-export { handleSamsAuthentication } from "./student-attendance-management-system";
-export {} from "./grading-management-system";
