+++ server/src/routes/index.route.ts
@@ -1,0 +1,11 @@
+import studentRouter from "./student.route";
+import collegeSubjectRouter from "./college-subject.route";
+import instructorRouter from "./instructor.route";
+import webhookRouter from './webhook.route'
+import authRouter from './auth.route'
+import samsRouter from './student-attendance-management-system/sams.route';
+import gmsRouter from './grading-management-system'
+import userRouter from './user.route'
+import strandRouter from './strand.route'
+
+export { userRouter, studentRouter, collegeSubjectRouter, instructorRouter, webhookRouter, authRouter, samsRouter, gmsRouter, strandRouter  };
