--- server/src/routes/index.ts
@@ -1,11 +1,0 @@
-import studentRouter from "./student";
-import collegeSubjectRouter from "./college-subject";
-import instructorRouter from "./instructor";
-import webhookRouter from './webhook'
-import authRouter from './auth'
-import samsRouter from './student-attendance-management-system';
-import gmsRouter from './grading-management-system'
-import userRouter from './user'
-import strandRouter from './strand'
-
-export { userRouter, studentRouter, collegeSubjectRouter, instructorRouter, webhookRouter, authRouter, samsRouter, gmsRouter, strandRouter  };
